// src/core/types/entities/dma/monitoring-status/List.ts
export type List = {
  dailyScheduleId: string
  unitNumber: string
  equipmentModel: string
  brand: string
  smuDue: string
  workOrder: string
  psType: string
  dateService: string
  shift: string
  isActive: string
  startDate: string
  endDate: string
  createdOn: string
  createdBy: string
  changedOn: string
  changedBy: string
  eFormId: string | null
  eFormKey: string | null
  status: string | null
  defectStatus: string | null
  isDownload?: string
}